requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pillow==10.3.0
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0